fixed warnings x4

+++ src/Headers/Game.h
@@ -28,7 +28,7 @@ public:
 
     void playCardFromHand(Player &player, Player &opponent, long unsigned int number_of_card);
 
-    static void playCardFromField(Player &player, Player &opponent, long unsigned int who_attack, long unsigned int target);
+    static void playCardFromField(Player &player, Player &opponent, long unsigned int who_attack, int target);
 
     void saveTheGame();
